requirements.txt
asgiref==3.5.2
autopep8==1.7.0
cffi==1.15.1
cryptography==37.0.4
Django==4.0.6
django-cors-headers==3.13.0
django-filter==22.1
django-rest-knox==4.2.0
djangorestframework==3.13.1
#psycopg2-binary==2.9.3
pycodestyle==2.9.1
pycparser==2.21
python-http-client==3.3.7
pytz==2022.1
sendgrid==6.9.7
sqlparse==0.4.2
starkbank-ecdsa==2.2.0
toml==0.10.2
tzdata==2022.1